chatbot/requirements.txt
numpy
streamlit
google-generativeai
python-dotenv
Pillow



# pip install google-generativeai  